Update cs205_final_exam.sh
As per the instructions commented in. I attempted to output file name, how many there are, their average hp, and their average attack. This was rather difficult in some ways. For the name I simply printed it for the sake of simplicity to output pokemon.dat, as that is the only file we will use for this program. For the number of pokemon we only had to count the lines. For the average hp and attack, I found it was easiest to loop through and add up the values for hp and attack, then divide them by the sum (which we've added up with each loop).