Update cs205_final_exam.sh
As per the instructions, I added a print line to give the file name. I then counted the lines to find how many pokemon there are. For hp and attack I only had to count add up the values in each column, then divide that by the sum of the columns.